Revert "Update for machine learning team. "
