+++ .gitignore
@@ -23,7 +23,6 @@ var/
 *.egg-info/
 .installed.cfg
 *.egg
-*.ipynb_checkpoints
 
 # PyInstaller
 #  Usually these files are written by a python script from a template
